services/auth-service/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS SDK
boto3==1.34.10
botocore==1.34.10

# PostgreSQL Database Driver
psycopg2-binary
psycopg2

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Data Validation
pydantic==2.5.2
pydantic[email]==2.5.2


# Logging
python-json-logger==2.0.7

# Environment Variables
python-dotenv==1.0.0

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Type Hints
typing-extensions==4.9.0

# Time & Date
python-dateutil==2.8.2
